+++ src/handler.rs
@@ -9,11 +9,13 @@ use crate::{
 };
 
 /// A request handler is an async function that accepts zero or more parameters that can be
-/// extracted from a request (i.e., [`impl FromRequest`](crate::FromRequest)) and returns a type
-/// that can be converted into an [`HttpResponse`] (that is, it impls the [`Responder`] trait).
+/// extracted from a request (i.e., [`impl FromRequest`]) and returns a type that can be converted
+/// into an [`HttpResponse`] (that is, it impls the [`Responder`] trait).
 ///
 /// If you got the error `the trait Handler<_, _, _> is not implemented`, then your function is not
-/// a valid handler. See [Request Handlers](https://actix.rs/docs/handlers/) for more information.
+/// a valid handler. See <https://actix.rs/docs/handlers> for more information.
+///
+/// [`impl FromRequest`]: crate::FromRequest
 pub trait Handler<T, R>: Clone + 'static
 where
     R: Future,
@@ -22,7 +24,7 @@ where
     fn call(&self, param: T) -> R;
 }
 
-pub fn handler_service<F, T, R>(handler: F) -> BoxedHttpServiceFactory
+pub(crate) fn handler_service<F, T, R>(handler: F) -> BoxedHttpServiceFactory
 where
     F: Handler<T, R>,
     T: FromRequest,
@@ -36,8 +38,9 @@ where
 
         async move {
             let (req, mut payload) = req.into_parts();
+
             let res = match T::from_request(&req, &mut payload).await {
-                Err(err) => HttpResponse::from_error(err).map_into_boxed_body(),
+                Err(err) => HttpResponse::from_error(err),
 
                 Ok(data) => handler
                     .call(data)
@@ -51,7 +54,14 @@ where
     }))
 }
 
-/// FromRequest trait impl for tuples
+/// Generates a [`Handler`] trait impl for N-ary functions where N is specified with a sequence of
+/// space separated type parameters.
+///
+/// # Examples
+/// ```ignore
+/// factory_tuple! {}         // implements Handler for types: fn() -> Res
+/// factory_tuple! { A B C }  // implements Handler for types: fn(A, B, C) -> Res
+/// ```
 macro_rules! factory_tuple ({ $($param:ident)* } => {
     impl<Func, $($param,)* Res> Handler<($($param,)*), Res> for Func
     where Func: Fn($($param),*) -> Res + Clone + 'static,